--- a/numpy/core/tests/test_records.py
+++ b/numpy/core/tests/test_records.py
@@ -437,6 +437,14 @@ class TestRecord(object):
         arr = np.zeros((3,), dtype=[('x', int), ('y', int)])
         assert_raises(ValueError, lambda: arr[['nofield']])
 
+    def test_frommarrays_using_subarrays(self):
+        # https://github.com/numpy/numpy/issues/4806
+        arrays = [
+            np.arange(10),
+            np.ones(10, dtype=[('a', '<u2'), ('b', '<f4')]),
+        ]
+        arr = np.rec.fromarrays(arrays) # ValueError?
+
 
 def test_find_duplicate():
     l1 = [1, 2, 3, 4, 5, 6]