* skip redundant check (all input arrays are at that point instances of _ndarray_)
* use _dtype_ instead of _dtype.str_

There is an improvement of ~3% 

```
· Running 14 total benchmarks (2 commits * 1 environments * 7 benchmarks)
[  0.00%] · For numpy commit 18435496 <master> (round 1/2):
[  0.00%] ·· Building for virtualenv-py3.6-six..
[  0.00%] ·· Benchmarking virtualenv-py3.6-six
[  3.57%] ··· Running (bench_records.Records.time_fromarrays_formats_as_list--).......
[ 25.00%] · For numpy commit 2c402b5c <danielhrisca-fromarrays> (round 1/2):
[ 25.00%] ·· Building for virtualenv-py3.6-six..
[ 25.00%] ·· Benchmarking virtualenv-py3.6-six
[ 28.57%] ··· Running (bench_records.Records.time_fromarrays_formats_as_list--).......
[ 50.00%] · For numpy commit 2c402b5c <danielhrisca-fromarrays> (round 2/2):
[ 50.00%] ·· Benchmarking virtualenv-py3.6-six
[ 53.57%] ··· bench_records.Records.time_fromarrays_formats_as_list                                                                 83.4±0.4ms
[ 57.14%] ··· bench_records.Records.time_fromarrays_formats_as_string                                                                108±0.3ms
[ 60.71%] ··· bench_records.Records.time_fromarrays_w_dtype                                                                         66.0±0.4ms
[ 64.29%] ··· bench_records.Records.time_fromarrays_wo_dtype                                                                        86.9±0.5ms
[ 67.86%] ··· bench_records.Records.time_fromstring_formats_as_list                                                                 16.6±0.1ms
[ 71.43%] ··· bench_records.Records.time_fromstring_formats_as_string                                                                 42.3±1ms
[ 75.00%] ··· bench_records.Records.time_fromstring_w_dtype                                                                        5.53±0.04μs
[ 75.00%] · For numpy commit 18435496 <master> (round 2/2):
[ 75.00%] ·· Building for virtualenv-py3.6-six..
[ 75.00%] ·· Benchmarking virtualenv-py3.6-six
[ 78.57%] ··· bench_records.Records.time_fromarrays_formats_as_list                                                                 85.5±0.2ms
[ 82.14%] ··· bench_records.Records.time_fromarrays_formats_as_string                                                                110±0.3ms
[ 85.71%] ··· bench_records.Records.time_fromarrays_w_dtype                                                                         68.5±0.2ms
[ 89.29%] ··· bench_records.Records.time_fromarrays_wo_dtype                                                                        89.9±0.1ms
[ 92.86%] ··· bench_records.Records.time_fromstring_formats_as_list                                                                 16.5±0.1ms
[ 96.43%] ··· bench_records.Records.time_fromstring_formats_as_string                                                               42.1±0.9ms
[100.00%] ··· bench_records.Records.time_fromstring_w_dtype                                                                        5.66±0.07μs
```

<!-- Please be sure you are following the instructions in the dev guidelines
http://www.numpy.org/devdocs/dev/gitwash/development_workflow.html
-->

<!-- We'd appreciate it if your commit message is properly formatted
http://www.numpy.org/devdocs/dev/gitwash/development_workflow.html#writing-the-commit-message
-->