@danielhrisca 2019-02-01
@tylerjereddy does this need a restart on azure?