@tylerjereddy 2019-02-01
@danielhrisca No, most likely you need to rebase on the latest master branch and force push so that your changes are applied on top of the recently-fixed Azure config file.