@mattip 2019-03-10
@eric-wieser the merge message should change too, since this is no longer a performance enhancement