@eric-wieser 2019-03-15
Thanks @danielhrisca, and sorry for taking almost 20 days to remember to merge this. Please ping next time!